+++ cmd/attach.go
@@ -213,15 +213,16 @@ func runAttach(cmd *cobra.Command) error {
 	containerGID := workspaceYAML.Spec.Container.GID
 
 	containerID, err := runtime.StartWorkspace(context.Background(), operators.StartOptions{
-		ImageName:     imageName,
-		WorkspaceName: workspaceName,
-		ContainerName: containerName,
-		AppName:       appName,
-		EcosystemName: ecosystemName,
-		DomainName:    domainName,
-		AppPath:       mountPath,
-		UID:           containerUID,
-		GID:           containerGID,
+		ImageName:          imageName,
+		WorkspaceName:      workspaceName,
+		ContainerName:      containerName,
+		AppName:            appName,
+		EcosystemName:      ecosystemName,
+		DomainName:         domainName,
+		AppPath:            mountPath,
+		UID:                containerUID,
+		GID:                containerGID,
+		SSHAgentForwarding: workspace.SSHAgentForwarding,
 	})
 	if err != nil {
 		return fmt.Errorf("failed to start workspace: %w", err)